+++ src/connectionimpl.cpp
@@ -146,7 +146,7 @@ uint64_t ConnectionImpl::parse(const Buffer &buffer)
                 // have the initial bytes of the header, we already know how much 
                 // data we need for the next frame, otherwise we need at least 7
                 // bytes for processing the header of the next frame
-                _expected = receivedFrame.header() ? receivedFrame.totalSize() : 7;
+                _expected = receivedFrame.header() ? (uint32_t)receivedFrame.totalSize() : 7;
 
                 // we're ready for now
                 return processed;
